+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
+- Color coverage statement result, green if 100%, magenta otherwise
 - Add usage to cmd/gocolor
 
 ## [0.3.0] - 2022-03-24